Last data added after closing the form. Migration index updated.
