+++ packages/lesswrong/server/manualMigrations/index.ts
@@ -118,3 +118,4 @@ import './2023-01-13-createViewUpdater'
 import './2023-01-18-user-mentions'
 import './2023-02-10-fillCreatedAtForDebouncerEvents'
 import './2023-03-16-createUserActivities.ts'
+import './2023-04-09-importACXMeetups'
